requirements.txt
streamlit
pandas
numpy
matplotlib
python-dotenv
litellm
seaborn
